Move _fixSocket from bitmessagemain to helper_startup,
so it can be used in test_networkgroup.
